pyproject.toml
[tool.poetry]
name = "pyjoibot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
python-decouple = "^3.4"
loguru = "^0.5.3"
toml = "^0.10.2"
python-slugify = "^4.0.1"
sentry-sdk = "^1.0.0"
httpx = "^0.17.1"
pytz = "^2021.1"
python-telegram-bot = "^13.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
